requirements.txt
pandas~=2.0.3
numpy~=1.24.4
scipy~=1.10.1
scikit-learn~=1.3.2
torch-geometric~=2.6.1